Move correction temperature constant
