+++ services/cartographer/AGENTS.md
@@ -8,3 +8,17 @@ The **cartographer** service is responsible for updating the map. It calls the a
 
 Edits here should maintain strict validation rules so invalid map structures do not propagate to the rest of the game.
 Both null and undefined optional fields in AI responses should be sanitized to undefined and treated as undefined down the line.
+
+### Gemini API call guidelines
+
+- Wrap Gemini requests in `retryAiCall` so temporary errors trigger retries.
+- Call `addProgressSymbol` before dispatching to give feedback.
+- All map update requests must use `dispatchAIRequest`.
+
+### Variable naming guidelines
+
+- `prompt` – string sent as the user content of the AI request.
+- `systemInstruction` – accompanying system instruction string.
+- `aiResponse` – raw value returned from the AI call.
+- `parsedResult` – JSON-parsed form of `aiResponse` when needed.
+- `validated*` – results after type validation, e.g. `validatedChanges`.
