+++ services/corrections/AGENTS.md
@@ -1,6 +1,6 @@
 The **corrections** helpers attempt to fix malformed AI responses. They are used when JSON fails validation or when map nodes require additional information.
 
-* `base.ts` – shared helpers for calling the correction models. Provides `callCorrectionAI` and `callMinimalCorrectionAI` wrappers that rely on `dispatchAIRequest`.
+* `constants.ts` – defines `CORRECTION_TEMPERATURE` used when dispatching AI requests.
 * `character.ts`, `item.ts`, `name.ts` – small utilities that request corrected data for specific entity types.
 * `map.ts` – sophisticated routines that repair or refine `AIMapUpdatePayload` results.
 * `dialogue.ts` – fixes issues with conversation summaries and turn outputs.
@@ -8,6 +8,12 @@ The **corrections** helpers attempt to fix malformed AI responses. They are used
 
 Use these utilities to keep the game running smoothly when the main models return inconsistent data.
 Both null and undefined optional fields in AI responses should be sanitized to undefined and treated as undefined down the line.
+
+### Gemini API call guidelines
+
+- Wrap each Gemini request in `retryAiCall` so transient API failures are retried automatically.
+- Call `addProgressSymbol` before dispatching to display progress feedback.
+- Route all requests through `dispatchAIRequest` specifying the desired model names and label.
 
 ### Variable naming guidelines
