+++ services/corrections/index.ts
@@ -2,10 +2,12 @@
  * @file services/corrections/index.ts
  * @description Central exports for correction helpers.
  */
-export * from './base';
 export * from './character';
 export * from './item';
 export * from './map';
+export * from './placeDetails';
+export * from './edgeFixes';
+export * from './hierarchyUpgrade';
 export * from './name';
 export * from './dialogue';
 export * from './inventory';
