+++ services/corrections/inventory.ts
@@ -7,11 +7,16 @@ import {
   MAX_RETRIES,
   VALID_ITEM_TYPES_STRING,
   PLAYER_HOLDER_ID,
+  AUXILIARY_MODEL_NAME,
+  GEMINI_MODEL_NAME,
 } from '../../constants';
-import { callCorrectionAI } from './base';
+import { CORRECTION_TEMPERATURE } from '../../constants';
+import { dispatchAIRequest } from '../modelDispatcher';
+import { addProgressSymbol } from '../../utils/loadingProgress';
 import { isApiConfigured } from '../apiClient';
 import { retryAiCall } from '../../utils/retry';
 import { parseInventoryResponse } from '../inventory/responseParser';
+import { extractJsonFromFence, safeParseJson } from '../../utils/jsonUtils';
 
 const VALID_ACTIONS = ['gain', 'destroy', 'update', 'put', 'give', 'take'] as const;
 const VALID_ACTIONS_STRING = VALID_ACTIONS.map(a => `"${a}"`).join(' | ');
@@ -61,7 +66,16 @@ Task: Provide ONLY the corrected JSON array of ItemChange objects.`;
 
   return retryAiCall<ItemChange[]>(async attempt => {
     try {
-      const aiResponse = await callCorrectionAI<ItemChange[]>(prompt, systemInstruction);
+      addProgressSymbol('●');
+      const { response } = await dispatchAIRequest({
+        modelNames: [AUXILIARY_MODEL_NAME, GEMINI_MODEL_NAME],
+        prompt,
+        systemInstruction,
+        responseMimeType: 'application/json',
+        temperature: CORRECTION_TEMPERATURE,
+        label: 'Corrections',
+      });
+      const aiResponse = safeParseJson<ItemChange[]>(extractJsonFromFence(response.text ?? ''));
       const parsedResult = aiResponse ? parseInventoryResponse(JSON.stringify(aiResponse)) : null;
       const validatedChanges = parsedResult ? parsedResult.itemChanges : null;
       if (validatedChanges) {
