+++ services/inventory/AGENTS.md
@@ -8,3 +8,17 @@ The **inventory** service coordinates small item management requests to the AI. 
 
 Maintain compatibility between the prompt builder and parser when adjusting this service.
 Both null and undefined optional fields in AI responses should be sanitized to undefined and treated as undefined down the line.
+
+### Gemini API call guidelines
+
+- Use `retryAiCall` to retry Gemini requests when needed.
+- Add a progress symbol via `addProgressSymbol` before dispatching.
+- Requests must go through `dispatchAIRequest`.
+
+### Variable naming guidelines
+
+- `prompt` – string sent as the user content of the AI request.
+- `systemInstruction` – accompanying system instruction string.
+- `aiResponse` – raw value returned from the AI call.
+- `parsedResult` – JSON-parsed form of `aiResponse` when needed.
+- `validated*` – results after type validation, e.g. `validatedChanges`.
