+++ services/storyteller/AGENTS.md
@@ -8,3 +8,17 @@ The **storyteller** service encapsulates all communication with the main narrati
 
 When modifying this service keep the prompt format and returned interfaces aligned with code in `useGameLogic` and the AI prompts.
 Both null and undefined optional fields in AI responses should be sanitized to undefined and treated as undefined down the line.
+
+### Gemini API call guidelines
+
+- Wrap Gemini requests in `retryAiCall` so transient failures are retried.
+- Use `addProgressSymbol` before dispatching to provide progress updates.
+- Always send requests via `dispatchAIRequest`.
+
+### Variable naming guidelines
+
+- `prompt` – string sent as the user content of the AI request.
+- `systemInstruction` – accompanying system instruction string.
+- `aiResponse` – raw value returned from the AI call.
+- `parsedResult` – JSON-parsed form of `aiResponse` when needed.
+- `validated*` – results after type validation, e.g. `validatedChanges`.
